dmpAPi/requirements.txt
asyncio
threaded
kafka-python
elasticsearch[async]
pymssql
python-dotenv
fastapi
uvicorn[standard]